Implementation of Automated Sum-Update
This is a naiive implementation of javascript triggered update fields.
Any input may have a "data-updates" attribute that contains a css
selector. Whenever such an input changes all inputs described by the
selector are changed correspondingly. The new content of these inputs
will be the sum of all elements that have a "data-updates" attribute
selecting it.
Not, that this update works on input fields only!
The new procedure renders the old widgets.js file deprecated. Thus it
can be removed. Furthermore, it is necessary to update to prototype
version 1.7. In order to not interfer with the automated grails update
process the prototype library has been added as an additional library.
While grails comes with anything less than 1.7 this file has to be
added. Once grails updates to 1.7 it can be removed.