Update to use Prototype 1.7
So far the application simply included Script.aculo.us, coming with
Prototype 1.6.1. While the Script.aculo.us library is not yet needed,
the Prototype 1.6.1 version is outdated. Due to reasons described in the
previous commit, the prototype version has to be changed to 1.7. Hence,
Script.aculo.us support is dropped for now and the new Prototype version
selected manually.