Rework of the UI including automated Updates
This is a general rework of the UI making use of the features
implemented in the previous commits. Due to concurrent work on all of
these features it was impossible to split it up into several commits.
1. Change of the views to use the new layout files. This includes a
sever rework of class names in use.
2. Rename of the "ausruestung" section into "behinderung". This change
was already included in the css rework. It is due to the fact that the
primary reason of this section is NOT to select proper equipment. Such a
section may be added later. Rather, it aims at identifying the proper
bonus imposed upon the bodily talents by equipment and damage.
3. Implementation of a naiive Number-With-Bonus field. These fields
allow users to update a number field, while a bonus field is updated by
other components. An additional field shows the sum of the number and
the bonus, thus representing the current effective value. The naiive
implementation uses the previously implemented automated update
function. Styles for these fields were provided be the previous css
udpate.