Change of the AutoUpdater
Until now any item was able to automatically update one selection of
items with its value. The new version allows specification of a number
of selectors where each selector can have a unique JavaScript function
that evaluates the added value at runtime. This does not necessarily
have to have anything to do with the actual element value. Furthermore,
every updater may specify a message that again is either a String with a
constant message or a JavaScript function.