Domain Rework
This a rework of the domain. First of all the Vorteilsexklusion has been
removed. The same functionality can be acquired much more easily using
the voraussetzungsScript attribute of the Vorteil domain. Second, the
Spezialisierung class has been split up into Talentspezialisierung and
Talenttypspezialisierung, thus allow definition of spezialisierungen on
a Talenttyp wide basis.